pyproject.toml
[tool.poetry]
name = "lesson-summarizer"
version = "0.1.0"
description = "Summarization pipeline for YouTube/audio/text"
authors = ["Martin Berns"]
readme = "README.md"
packages = [{ include = "lesson_sumarizer", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
streamlit = "^1.39.0"
python-dotenv = "^1.0.1"
pydantic = "^2.8.2"
requests = "^2.32.3"
numpy = "^2.0.1"
sentence-transformers = "^3.0.1"
yt-dlp = "^2025.1.1"
reportlab = "^4.4.7"
python-docx = "^1.2.0"
pyyaml = "^6.0.3"
google-genai = "^1.56.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.9"

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"